pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quarterly-financials"
version = "0.1.0"
description = "Application pour comparer les résultats trimestriels de sociétés e-commerce et crypto"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "sqlalchemy>=2.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
    "pdfplumber>=0.10",
    "pandas>=2.1",
    "typer>=0.9",
    "rich>=13.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
qf = "src.cli.commands:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
